rremoved classification benchmark changes - move to branch
