+++ machine_learning/benchmarks/circuit_qnn_classifier_fit_benchmark.py
@@ -14,6 +14,7 @@
 from itertools import product
 from timeit import timeit
 
+import numpy as np
 from qiskit import QuantumCircuit
 from qiskit.circuit.library import ZZFeatureMap, RealAmplitudes
 from qiskit.algorithms.optimizers import COBYLA, NELDER_MEAD, L_BFGS_B
@@ -45,9 +46,9 @@ class CircuitQnnFitClassifierBenchmarks(BaseClassifierBenchmark):
 
     def setup(self, dataset, quantum_instance_name, optimizer_name):
         """setup"""
-        self.X = self.datasets[dataset]["features"]
+        self.X = self.datasets[dataset]
         num_inputs = len(self.X[0])
-        self.y01 = self.datasets[dataset]["labels"]
+        self.y01 = 1 * (np.sum(self.X, axis=1) >= 0)  # in { 0,  1}
 
         # construct feature map
         feature_map = ZZFeatureMap(num_inputs)
