+++ machine_learning/benchmarks/vqc_benchmark.py
@@ -33,10 +33,10 @@ class VqcBenchmarks(BaseClassifierBenchmark):
 
     def setup(self, dataset, quantum_instance_name):
         """setup"""
-        self.X = self.datasets[dataset]["features"]
+        self.X = self.datasets[dataset]
         num_inputs = len(self.X[0])
         num_samples = len(self.X)
-        y01 = self.datasets[dataset]["labels"]
+        y01 = 1 * (np.sum(self.X, axis=1) >= 0)  # in { 0,  1}
         self.y_one_hot = np.zeros((num_samples, 2))
         for i in range(num_samples):
             self.y_one_hot[i, y01[i]] = 1
